Move: Movendo codigo para uma melhor org
