+++ cmd/create.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -12,6 +11,9 @@ import (
 	"strings"
 	"time"
 
+	"github.com/badtuxx/girus-cli/internal/helpers"
+	"github.com/badtuxx/girus-cli/internal/k8s"
+	"github.com/badtuxx/girus-cli/internal/lab"
 	"github.com/badtuxx/girus-cli/internal/templates"
 	"github.com/schollz/progressbar/v3"
 	"github.com/spf13/cobra"
@@ -25,388 +27,6 @@ var (
 	skipPortForward bool
 	skipBrowser     bool
 )
-
-// waitForPodsReady espera até que os pods do Girus (backend e frontend) estejam prontos
-func waitForPodsReady(namespace string, timeout time.Duration) error {
-	fmt.Println("\nAguardando os pods do Girus inicializarem...")
-
-	start := time.Now()
-	bar := progressbar.NewOptions(100,
-		progressbar.OptionSetDescription("Inicializando Girus..."),
-		progressbar.OptionSetWidth(80),
-		progressbar.OptionShowBytes(false),
-		progressbar.OptionSetPredictTime(false),
-		progressbar.OptionThrottle(65*time.Millisecond),
-		progressbar.OptionSetRenderBlankState(true),
-		progressbar.OptionSpinnerType(14),
-		progressbar.OptionFullWidth(),
-	)
-
-	backendReady := false
-	frontendReady := false
-	backendMessage := ""
-	frontendMessage := ""
-
-	for {
-		if time.Since(start) > timeout {
-			bar.Finish()
-			fmt.Println("\nStatus atual dos componentes:")
-			if backendReady {
-				fmt.Printf("✅ Backend: Pronto\n")
-			} else {
-				fmt.Printf("❌ Backend: %s\n", backendMessage)
-			}
-			if frontendReady {
-				fmt.Printf("✅ Frontend: Pronto\n")
-			} else {
-				fmt.Printf("❌ Frontend: %s\n", frontendMessage)
-			}
-			return fmt.Errorf("timeout ao esperar pelos pods do Girus (5 minutos)")
-		}
-
-		// Verificar o backend
-		if !backendReady {
-			br, msg, err := getPodStatus(namespace, "app=girus-backend")
-			if err == nil {
-				backendReady = br
-				backendMessage = msg
-			}
-		}
-
-		// Verificar o frontend
-		if !frontendReady {
-			fr, msg, err := getPodStatus(namespace, "app=girus-frontend")
-			if err == nil {
-				frontendReady = fr
-				frontendMessage = msg
-			}
-		}
-
-		// Se ambos estiverem prontos, vamos verificar a conectividade
-		if backendReady && frontendReady {
-			// Verificar se conseguimos acessar a API
-			healthy, err := checkHealthEndpoint()
-			if err != nil || !healthy {
-				// Ainda não está respondendo, vamos continuar tentando
-				bar.Add(1)
-				time.Sleep(500 * time.Millisecond)
-				continue
-			}
-
-			bar.Finish()
-			fmt.Println("\n✅ Backend: Pronto")
-			fmt.Println("✅ Frontend: Pronto")
-			fmt.Println("✅ Aplicação: Respondendo")
-			return nil
-		}
-
-		bar.Add(1)
-		time.Sleep(500 * time.Millisecond)
-	}
-}
-
-// getPodStatus verifica o status de um pod e retorna uma mensagem descritiva
-func getPodStatus(namespace, selector string) (bool, string, error) {
-	// Verificar se o pod existe
-	cmd := exec.Command("kubectl", "get", "pods", "-n", namespace, "-l", selector, "-o", "jsonpath={.items[0].metadata.name}")
-	var out bytes.Buffer
-	cmd.Stdout = &out
-
-	err := cmd.Run()
-	if err != nil {
-		return false, "Pod não encontrado", err
-	}
-
-	podName := strings.TrimSpace(out.String())
-	if podName == "" {
-		return false, "Pod ainda não criado", nil
-	}
-
-	// Verificar a fase atual do pod
-	phaseCmd := exec.Command("kubectl", "get", "pod", podName, "-n", namespace, "-o", "jsonpath={.status.phase}")
-	var phaseOut bytes.Buffer
-	phaseCmd.Stdout = &phaseOut
-
-	err = phaseCmd.Run()
-	if err != nil {
-		return false, "Erro ao verificar status", err
-	}
-
-	phase := strings.TrimSpace(phaseOut.String())
-	if phase != "Running" {
-		return false, fmt.Sprintf("Status: %s", phase), nil
-	}
-
-	// Verificar se todos os containers estão prontos
-	readyCmd := exec.Command("kubectl", "get", "pod", podName, "-n", namespace, "-o", "jsonpath={.status.conditions[?(@.type=='Ready')].status}")
-	var readyOut bytes.Buffer
-	readyCmd.Stdout = &readyOut
-
-	err = readyCmd.Run()
-	if err != nil {
-		return false, "Erro ao verificar prontidão", err
-	}
-
-	readyStatus := strings.TrimSpace(readyOut.String())
-	if readyStatus != "True" {
-		return false, "Containers inicializando", nil
-	}
-
-	return true, "Pronto", nil
-}
-
-// checkHealthEndpoint verifica se a aplicação está respondendo ao endpoint de saúde
-func checkHealthEndpoint() (bool, error) {
-	// Verificar o mapeamento de porta do serviço
-	cmd := exec.Command("kubectl", "get", "svc", "-n", "girus", "girus-backend", "-o", "jsonpath={.spec.ports[0].nodePort}")
-	var out bytes.Buffer
-	cmd.Stdout = &out
-
-	err := cmd.Run()
-	if err != nil {
-		// Tentar verificar diretamente o endpoint interno se não encontrarmos o NodePort
-		healthCmd := exec.Command("kubectl", "exec", "-n", "girus", "deploy/girus-backend", "--", "wget", "-q", "-O-", "-T", "2", "http://localhost:8080/api/v1/health")
-		return healthCmd.Run() == nil, nil
-	}
-
-	nodePort := strings.TrimSpace(out.String())
-	if nodePort == "" {
-		// Porta não encontrada, tentar verificar o serviço internamente
-		healthCmd := exec.Command("kubectl", "exec", "-n", "girus", "deploy/girus-backend", "--", "wget", "-q", "-O-", "-T", "2", "http://localhost:8080/api/v1/health")
-		return healthCmd.Run() == nil, nil
-	}
-
-	// Tentar acessar via NodePort
-	healthCmd := exec.Command("curl", "-s", "-o", "/dev/null", "-w", "%{http_code}", fmt.Sprintf("http://localhost:%s/api/v1/health", nodePort))
-	var healthOut bytes.Buffer
-	healthCmd.Stdout = &healthOut
-
-	err = healthCmd.Run()
-	if err != nil {
-		return false, err
-	}
-
-	statusCode := strings.TrimSpace(healthOut.String())
-	return statusCode == "200", nil
-}
-
-// setupPortForward configura port-forward para os serviços do Girus
-func setupPortForward(namespace string) error {
-	// Matar todos os processos de port-forward relacionados ao Girus para começar limpo
-	fmt.Println("   Limpando port-forwards existentes...")
-	exec.Command("bash", "-c", "pkill -f 'kubectl.*port-forward.*girus' || true").Run()
-	time.Sleep(1 * time.Second)
-
-	// Port-forward do backend em background
-	fmt.Println("   Configurando port-forward para o backend (8080)...")
-	backendCmd := fmt.Sprintf("kubectl port-forward -n %s svc/girus-backend 8080:8080 --address 0.0.0.0 > /dev/null 2>&1 &", namespace)
-	err := exec.Command("bash", "-c", backendCmd).Run()
-	if err != nil {
-		return fmt.Errorf("erro ao iniciar port-forward do backend: %v", err)
-	}
-
-	// Verificar conectividade do backend
-	fmt.Println("   Verificando conectividade do backend...")
-	backendOK := false
-	for i := 0; i < 5; i++ {
-		healthCmd := exec.Command("curl", "-s", "--max-time", "2", "http://localhost:8080/api/v1/health")
-		if healthCmd.Run() == nil {
-			backendOK = true
-			break
-		}
-		if i < 4 {
-			fmt.Println("   Tentativa", i+1, "falhou, aguardando...")
-			time.Sleep(1 * time.Second)
-		}
-	}
-
-	if !backendOK {
-		return fmt.Errorf("não foi possível conectar ao backend")
-	}
-
-	fmt.Println("   ✅ Backend conectado com sucesso!")
-
-	// ------------------------------------------------------------------------
-	// Port-forward do frontend - ABORDAGEM MAIS SIMPLES E DIRETA
-	// ------------------------------------------------------------------------
-	fmt.Println("   Configurando port-forward para o frontend (8000)...")
-
-	// Método 1: Execução direta via bash para o frontend
-	frontendSuccess := false
-
-	// Criar um script temporário para garantir execução correta
-	scriptContent := `#!/bin/bash
-# Mata qualquer processo existente na porta 8000
-kill $(lsof -t -i:8000) 2>/dev/null || true
-sleep 1
-# Inicia o port-forward
-nohup kubectl port-forward -n NAMESPACE svc/girus-frontend 8000:80 --address 0.0.0.0 > /dev/null 2>&1 &
-echo $!  # Retorna o PID
-`
-
-	// Substituir NAMESPACE pelo namespace real
-	scriptContent = strings.Replace(scriptContent, "NAMESPACE", namespace, 1)
-
-	// Salvar em arquivo temporário
-	tmpFile := filepath.Join(os.TempDir(), "girus_frontend_portforward.sh")
-	os.WriteFile(tmpFile, []byte(scriptContent), 0755)
-	defer os.Remove(tmpFile)
-
-	// Executar o script
-	fmt.Println("   Iniciando port-forward via script auxiliar...")
-	cmdOutput, err := exec.Command("bash", tmpFile).Output()
-	if err == nil {
-		pid := strings.TrimSpace(string(cmdOutput))
-		fmt.Println("   Port-forward iniciado com PID:", pid)
-
-		// Aguardar o port-forward inicializar
-		time.Sleep(2 * time.Second)
-
-		// Verificar conectividade
-		for i := 0; i < 5; i++ {
-			checkCmd := exec.Command("curl", "-s", "--max-time", "2", "-o", "/dev/null", "-w", "%{http_code}", "http://localhost:8000")
-			var out bytes.Buffer
-			checkCmd.Stdout = &out
-
-			if err := checkCmd.Run(); err == nil {
-				statusCode := strings.TrimSpace(out.String())
-				if statusCode == "200" || statusCode == "301" || statusCode == "302" {
-					frontendSuccess = true
-					break
-				}
-			}
-
-			fmt.Println("   Verificação", i+1, "falhou, aguardando...")
-			time.Sleep(2 * time.Second)
-		}
-	}
-
-	// Se falhou, tentar um método alternativo como último recurso
-	if !frontendSuccess {
-		fmt.Println("   ⚠️ Tentando método alternativo direto...")
-
-		// Método direto: executar o comando diretamente
-		cmd := exec.Command("kubectl", "port-forward", "-n", namespace, "svc/girus-frontend", "8000:80", "--address", "0.0.0.0")
-
-		// Redirecionar saída para /dev/null
-		devNull, _ := os.Open(os.DevNull)
-		defer devNull.Close()
-		cmd.Stdout = devNull
-		cmd.Stderr = devNull
-
-		// Iniciar em background - compatível com múltiplos sistemas operacionais
-		startBackgroundCmd(cmd)
-
-		// Verificar conectividade
-		time.Sleep(3 * time.Second)
-		for i := 0; i < 3; i++ {
-			checkCmd := exec.Command("curl", "-s", "--max-time", "2", "-o", "/dev/null", "-w", "%{http_code}", "http://localhost:8000")
-			var out bytes.Buffer
-			checkCmd.Stdout = &out
-
-			if err := checkCmd.Run(); err == nil {
-				statusCode := strings.TrimSpace(out.String())
-				if statusCode == "200" || statusCode == "301" || statusCode == "302" {
-					frontendSuccess = true
-					break
-				}
-			}
-			time.Sleep(1 * time.Second)
-		}
-	}
-
-	// Último recurso - método absolutamente direto com deployment em vez de service
-	if !frontendSuccess {
-		fmt.Println("   🔄 Último recurso: port-forward ao deployment...")
-		// Método com deployment em vez de service, que pode ser mais estável
-		finalCmd := fmt.Sprintf("kubectl port-forward -n %s deployment/girus-frontend 8000:80 --address 0.0.0.0 > /dev/null 2>&1 &", namespace)
-		exec.Command("bash", "-c", finalCmd).Run()
-
-		// Verificação final
-		time.Sleep(3 * time.Second)
-		checkCmd := exec.Command("curl", "-s", "--max-time", "2", "-o", "/dev/null", "-w", "%{http_code}", "http://localhost:8000")
-		var out bytes.Buffer
-		checkCmd.Stdout = &out
-
-		if checkCmd.Run() == nil {
-			statusCode := strings.TrimSpace(out.String())
-			if statusCode == "200" || statusCode == "301" || statusCode == "302" {
-				frontendSuccess = true
-			}
-		}
-	}
-
-	// Verificar status final e retornar
-	if !frontendSuccess {
-		return fmt.Errorf("não foi possível estabelecer port-forward para o frontend após múltiplas tentativas")
-	}
-
-	fmt.Println("   ✅ Frontend conectado com sucesso!")
-	return nil
-}
-
-// startBackgroundCmd inicia um comando em segundo plano de forma compatível com múltiplos sistemas operacionais
-func startBackgroundCmd(cmd *exec.Cmd) error {
-	// Iniciar o processo sem depender de atributos específicos da plataforma
-	// que podem não estar disponíveis em todas as implementações do Go
-
-	// Redirecionar saída e erro para /dev/null ou nul (Windows)
-	devNull, _ := os.Open(os.DevNull)
-	if devNull != nil {
-		cmd.Stdout = devNull
-		cmd.Stderr = devNull
-		defer devNull.Close()
-	}
-
-	// Iniciar o processo
-	err := cmd.Start()
-	if err != nil {
-		return err
-	}
-
-	// Registrar o PID para referência
-	if cmd.Process != nil {
-		homeDir, _ := os.UserHomeDir()
-		if homeDir != "" {
-			pidDir := filepath.Join(homeDir, ".girus")
-			os.MkdirAll(pidDir, 0755)
-			ioutil.WriteFile(filepath.Join(pidDir, "frontend.pid"),
-				[]byte(fmt.Sprintf("%d", cmd.Process.Pid)), 0644)
-		}
-
-		// Separar o processo do atual para evitar que seja terminado quando o processo pai terminar
-		// Isso é uma alternativa portable ao uso de Setpgid
-		go func() {
-			cmd.Process.Release()
-		}()
-	}
-
-	return nil
-}
-
-// portInUse verifica se uma porta está em uso
-func portInUse(port int) bool {
-	checkCmd := exec.Command("lsof", "-i", fmt.Sprintf(":%d", port))
-	return checkCmd.Run() == nil
-}
-
-// openBrowser abre o navegador com a URL especificada
-func openBrowser(url string) error {
-	var cmd *exec.Cmd
-
-	switch runtime.GOOS {
-	case "linux":
-		cmd = exec.Command("xdg-open", url)
-	case "darwin":
-		cmd = exec.Command("open", url)
-	case "windows":
-		cmd = exec.Command("cmd", "/c", "start", url)
-	default:
-		return fmt.Errorf("não foi possível abrir o navegador (sistema operacional não suportado)")
-	}
-
-	return cmd.Start()
-}
 
 var createCmd = &cobra.Command{
 	Use:   "create [subcommand]",
@@ -1301,7 +921,7 @@ Por padrão, o deployment embutido no binário é utilizado.`,
 		}
 
 		// Aguardar os pods do Girus ficarem prontos
-		if err := waitForPodsReady("girus", 5*time.Minute); err != nil {
+		if err := k8s.WaitForPodsReady("girus", 5*time.Minute); err != nil {
 			fmt.Fprintf(os.Stderr, "Aviso: %v\n", err)
 			fmt.Println("Recomenda-se verificar o estado dos pods com 'kubectl get pods -n girus'")
 		} else {
@@ -1314,7 +934,7 @@ Por padrão, o deployment embutido no binário é utilizado.`,
 		if !skipPortForward {
 			fmt.Print("\n🔌 Configurando acesso aos serviços do Girus... ")
 
-			if err := setupPortForward("girus"); err != nil {
+			if err := k8s.SetupPortForward("girus"); err != nil {
 				fmt.Println("⚠️")
 				fmt.Printf("Não foi possível configurar o acesso automático: %v\n", err)
 				fmt.Println("\nVocê pode tentar configurar manualmente com os comandos:")
@@ -1329,7 +949,7 @@ Por padrão, o deployment embutido no binário é utilizado.`,
 				// Abrir o navegador se não foi especificado para pular
 				if !skipBrowser {
 					fmt.Println("\n🌐 Abrindo navegador com o Girus...")
-					if err := openBrowser("http://localhost:8000"); err != nil {
+					if err := helpers.OpenBrowser("http://localhost:8000"); err != nil {
 						fmt.Printf("⚠️  Não foi possível abrir o navegador: %v\n", err)
 						fmt.Println("   Acesse manualmente: http://localhost:8000")
 					}
@@ -1371,7 +991,7 @@ var createLabCmd = &cobra.Command{
 		// Verificar qual modo estamos
 		if labFile != "" {
 			// Modo de adicionar template a partir de arquivo
-			addLabFromFile(labFile, verboseMode)
+			lab.AddLabFromFile(labFile, verboseMode)
 		} else {
 			fmt.Fprintf(os.Stderr, "Erro: Você deve especificar um arquivo de laboratório com a flag -f\n")
 			fmt.Println("\nExemplo:")
@@ -1380,434 +1000,6 @@ var createLabCmd = &cobra.Command{
 			os.Exit(1)
 		}
 	},
-}
-
-// addLabFromFile adiciona um novo template de laboratório a partir de um arquivo
-func addLabFromFile(labFile string, verboseMode bool) {
-	// Verificar se o arquivo existe
-	if _, err := os.Stat(labFile); os.IsNotExist(err) {
-		fmt.Fprintf(os.Stderr, "❌ Erro: arquivo '%s' não encontrado\n", labFile)
-		os.Exit(1)
-	}
-
-	fmt.Println("🔍 Verificando ambiente Girus...")
-
-	// Verificar se há um cluster Girus ativo
-	checkCmd := exec.Command("kubectl", "get", "namespace", "girus", "--no-headers", "--ignore-not-found")
-	checkOutput, err := checkCmd.Output()
-	if err != nil || !strings.Contains(string(checkOutput), "girus") {
-		fmt.Fprintf(os.Stderr, "❌ Nenhum cluster Girus ativo encontrado\n")
-		fmt.Println("   Use 'girus create cluster' para criar um cluster ou 'girus list clusters' para ver os disponíveis.")
-		os.Exit(1)
-	}
-
-	// Verificar o pod do backend (silenciosamente, só mostra mensagem em caso de erro)
-	backendCmd := exec.Command("kubectl", "get", "pods", "-n", "girus", "-l", "app=girus-backend", "-o", "jsonpath={.items[0].status.phase}")
-	backendOutput, err := backendCmd.Output()
-	if err != nil || string(backendOutput) != "Running" {
-		fmt.Fprintf(os.Stderr, "❌ O backend do Girus não está em execução\n")
-		fmt.Println("   Verifique o status dos pods com 'kubectl get pods -n girus'")
-		os.Exit(1)
-	}
-
-	// Ler o arquivo para verificar se é um ConfigMap válido
-	content, err := os.ReadFile(labFile)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "❌ Erro ao ler o arquivo '%s': %v\n", labFile, err)
-		os.Exit(1)
-	}
-
-	// Verificação simples se o arquivo parece ser um ConfigMap válido
-	fileContent := string(content)
-	if !strings.Contains(fileContent, "kind: ConfigMap") ||
-		!strings.Contains(fileContent, "app: girus-lab-template") {
-		fmt.Fprintf(os.Stderr, "❌ O arquivo não é um manifesto de laboratório válido\n")
-		fmt.Println("   O arquivo deve ser um ConfigMap com a label 'app: girus-lab-template'")
-		os.Exit(1)
-	}
-
-	// Verificar se está instalando o lab do Docker e se o Docker está disponível
-	if strings.Contains(fileContent, "docker-basics") {
-		fmt.Println("🐳 Detectado laboratório de Docker, verificando dependências...")
-
-		// Verificar se o Docker está instalado
-		dockerCmd := exec.Command("docker", "--version")
-		dockerInstalled := dockerCmd.Run() == nil
-
-		// Verificar se o serviço está rodando
-		dockerRunning := false
-		if dockerInstalled {
-			infoCmd := exec.Command("docker", "info")
-			dockerRunning = infoCmd.Run() == nil
-		}
-
-		if !dockerInstalled || !dockerRunning {
-			fmt.Println("⚠️  Aviso: Docker não está instalado ou não está em execução")
-			fmt.Println("   O laboratório de Docker será instalado, mas requer Docker para funcionar corretamente.")
-			fmt.Println("   Para instalar o Docker:")
-
-			if runtime.GOOS == "darwin" {
-				fmt.Println("\n   📦 macOS (via Colima):")
-				fmt.Println("      brew install colima docker")
-				fmt.Println("      colima start")
-			} else if runtime.GOOS == "linux" {
-				fmt.Println("\n   📦 Linux:")
-				fmt.Println("      curl -fsSL https://get.docker.com | bash")
-				fmt.Println("      sudo usermod -aG docker $USER")
-				fmt.Println("      sudo systemctl start docker")
-			} else {
-				fmt.Println("\n   📦 Visite: https://www.docker.com/products/docker-desktop")
-			}
-
-			fmt.Println("\n   Você deseja continuar com a instalação do template? [s/N]")
-			reader := bufio.NewReader(os.Stdin)
-			response, _ := reader.ReadString('\n')
-			response = strings.ToLower(strings.TrimSpace(response))
-
-			if response != "s" && response != "sim" && response != "y" && response != "yes" {
-				fmt.Println("Instalação cancelada.")
-				os.Exit(0)
-			}
-
-			fmt.Println("Continuando com a instalação do template Docker...")
-		} else {
-			fmt.Println("✅ Docker detectado e funcionando")
-		}
-	}
-
-	fmt.Printf("📦 Processando laboratório: %s\n", labFile)
-
-	// Aplicar o ConfigMap no cluster usando kubectl apply
-	if verboseMode {
-		fmt.Println("   Aplicando ConfigMap no cluster...")
-	}
-
-	// Aplicar o ConfigMap no cluster
-	if verboseMode {
-		// Executar normalmente mostrando o output
-		applyCmd := exec.Command("kubectl", "apply", "-f", labFile)
-		applyCmd.Stdout = os.Stdout
-		applyCmd.Stderr = os.Stderr
-		if err := applyCmd.Run(); err != nil {
-			fmt.Fprintf(os.Stderr, "❌ Erro ao aplicar o laboratório: %v\n", err)
-			os.Exit(1)
-		}
-	} else {
-		// Usar barra de progresso
-		bar := progressbar.NewOptions(100,
-			progressbar.OptionSetDescription("   Aplicando laboratório"),
-			progressbar.OptionSetWidth(80),
-			progressbar.OptionShowBytes(false),
-			progressbar.OptionSetPredictTime(false),
-			progressbar.OptionThrottle(65*time.Millisecond),
-			progressbar.OptionSetRenderBlankState(true),
-			progressbar.OptionSpinnerType(14),
-			progressbar.OptionFullWidth(),
-		)
-
-		// Executar comando sem mostrar saída
-		applyCmd := exec.Command("kubectl", "apply", "-f", labFile)
-		var stderr bytes.Buffer
-		applyCmd.Stderr = &stderr
-
-		// Iniciar o comando
-		err := applyCmd.Start()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "❌ Erro ao iniciar o comando: %v\n", err)
-			os.Exit(1)
-		}
-
-		// Atualizar a barra de progresso enquanto o comando está em execução
-		done := make(chan struct{})
-		go func() {
-			for {
-				select {
-				case <-done:
-					return
-				default:
-					bar.Add(1)
-					time.Sleep(50 * time.Millisecond)
-				}
-			}
-		}()
-
-		// Aguardar o final do comando
-		err = applyCmd.Wait()
-		close(done)
-		bar.Finish()
-
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "\n❌ Erro ao aplicar o laboratório: %v\n", err)
-			if verboseMode {
-				fmt.Fprintf(os.Stderr, "   Detalhes: %s\n", stderr.String())
-			}
-			os.Exit(1)
-		}
-	}
-
-	// Extrair o ID do lab (name) do arquivo YAML para mostrar na mensagem
-	var labID string
-	// Procurar pela linha 'name:' dentro do bloco lab.yaml:
-	labNameCmd := exec.Command("sh", "-c", fmt.Sprintf("grep -A10 'lab.yaml:' %s | grep 'name:' | head -1", labFile))
-	labNameOutput, err := labNameCmd.Output()
-	if err == nil {
-		nameLine := strings.TrimSpace(string(labNameOutput))
-		parts := strings.SplitN(nameLine, "name:", 2)
-		if len(parts) >= 2 {
-			labID = strings.TrimSpace(parts[1])
-		}
-	}
-
-	// Extrair também o título para exibição
-	var labTitle string
-	labTitleCmd := exec.Command("sh", "-c", fmt.Sprintf("grep -A10 'lab.yaml:' %s | grep 'title:' | head -1", labFile))
-	labTitleOutput, err := labTitleCmd.Output()
-	if err == nil {
-		titleLine := strings.TrimSpace(string(labTitleOutput))
-		parts := strings.SplitN(titleLine, "title:", 2)
-		if len(parts) >= 2 {
-			labTitle = strings.TrimSpace(parts[1])
-			labTitle = strings.Trim(labTitle, "\"'")
-		}
-	}
-
-	fmt.Println("\n🔄 Reiniciando backend para carregar o template...")
-
-	// O backend apenas carrega os templates na inicialização
-	if verboseMode {
-		// Mostrar o output da reinicialização
-		fmt.Println("   (O backend do Girus carrega os templates apenas na inicialização)")
-		restartCmd := exec.Command("kubectl", "rollout", "restart", "deployment/girus-backend", "-n", "girus")
-		restartCmd.Stdout = os.Stdout
-		restartCmd.Stderr = os.Stderr
-		if err := restartCmd.Run(); err != nil {
-			fmt.Fprintf(os.Stderr, "⚠️  Erro ao reiniciar o backend: %v\n", err)
-			fmt.Println("   O template foi aplicado, mas pode ser necessário reiniciar o backend manualmente:")
-			fmt.Println("   kubectl rollout restart deployment/girus-backend -n girus")
-		}
-
-		// Aguardar o reinício completar
-		fmt.Println("   Aguardando o reinício do backend completar...")
-		waitCmd := exec.Command("kubectl", "rollout", "status", "deployment/girus-backend", "-n", "girus", "--timeout=60s")
-		// Redirecionar saída para não exibir detalhes do rollout
-		var waitOutput bytes.Buffer
-		waitCmd.Stdout = &waitOutput
-		waitCmd.Stderr = &waitOutput
-
-		// Iniciar indicador de progresso simples
-		spinChars := []string{"⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"}
-		spinIdx := 0
-		done := make(chan struct{})
-		go func() {
-			for {
-				select {
-				case <-done:
-					return
-				default:
-					fmt.Printf("\r   %s Aguardando... ", spinChars[spinIdx])
-					spinIdx = (spinIdx + 1) % len(spinChars)
-					time.Sleep(100 * time.Millisecond)
-				}
-			}
-		}()
-
-		// Executar e aguardar
-		waitCmd.Run()
-		close(done)
-		fmt.Println("\r   ✅ Backend reiniciado com sucesso!            ")
-	} else {
-		// Usar barra de progresso
-		bar := progressbar.NewOptions(100,
-			progressbar.OptionSetDescription("   Reiniciando backend"),
-			progressbar.OptionSetWidth(80),
-			progressbar.OptionShowBytes(false),
-			progressbar.OptionSetPredictTime(false),
-			progressbar.OptionThrottle(65*time.Millisecond),
-			progressbar.OptionSetRenderBlankState(true),
-			progressbar.OptionSpinnerType(14),
-			progressbar.OptionFullWidth(),
-		)
-
-		// Reiniciar o deployment do backend
-		restartCmd := exec.Command("kubectl", "rollout", "restart", "deployment/girus-backend", "-n", "girus")
-		var stderr bytes.Buffer
-		restartCmd.Stderr = &stderr
-
-		err := restartCmd.Run()
-		if err != nil {
-			bar.Finish()
-			fmt.Fprintf(os.Stderr, "\n⚠️  Erro ao reiniciar o backend: %v\n", err)
-			if verboseMode {
-				fmt.Fprintf(os.Stderr, "   Detalhes: %s\n", stderr.String())
-			}
-			fmt.Println("   O template foi aplicado, mas pode ser necessário reiniciar o backend manualmente:")
-			fmt.Println("   kubectl rollout restart deployment/girus-backend -n girus")
-		} else {
-			// Aguardar o reinício completar
-			waitCmd := exec.Command("kubectl", "rollout", "status", "deployment/girus-backend", "-n", "girus", "--timeout=60s")
-
-			// Redirecionar saída para não exibir detalhes do rollout
-			var waitOutput bytes.Buffer
-			waitCmd.Stdout = &waitOutput
-			waitCmd.Stderr = &waitOutput
-
-			// Iniciar o comando
-			err = waitCmd.Start()
-			if err != nil {
-				bar.Finish()
-				fmt.Fprintf(os.Stderr, "\n⚠️  Erro ao verificar status do reinício: %v\n", err)
-			} else {
-				// Atualizar a barra de progresso enquanto o comando está em execução
-				done := make(chan struct{})
-				go func() {
-					for {
-						select {
-						case <-done:
-							return
-						default:
-							bar.Add(1)
-							time.Sleep(100 * time.Millisecond)
-						}
-					}
-				}()
-
-				// Aguardar o final do comando
-				waitCmd.Wait()
-				close(done)
-				fmt.Println("\r   ✅ Backend reiniciado com sucesso!            ")
-			}
-			bar.Finish()
-		}
-	}
-
-	// Aguardar mais alguns segundos para que o backend reinicie completamente
-	fmt.Println("   Aguardando inicialização completa...")
-	time.Sleep(3 * time.Second)
-
-	// Após reiniciar o backend, verificar se precisamos recriar o port-forward
-	portForwardStatus := checkPortForwardNeeded()
-
-	// Se port-forward é necessário, configurá-lo corretamente
-	if portForwardStatus {
-		fmt.Println("\n🔌 Reconfigurando port-forwards após reinício do backend...")
-
-		// Usar a função setupPortForward para garantir que ambos os serviços estejam acessíveis
-		err := setupPortForward("girus")
-		if err != nil {
-			fmt.Println("⚠️ Aviso:", err)
-			fmt.Println("   Para configurar manualmente, execute:")
-			fmt.Println("   kubectl port-forward -n girus svc/girus-backend 8080:8080 --address 0.0.0.0")
-			fmt.Println("   kubectl port-forward -n girus svc/girus-frontend 8000:80 --address 0.0.0.0")
-		} else {
-			fmt.Println("✅ Port-forwards configurados com sucesso!")
-			fmt.Println("   🔹 Backend: http://localhost:8080")
-			fmt.Println("   🔹 Frontend: http://localhost:8000")
-		}
-	} else {
-		// Verificar conexão com o frontend mesmo que o port-forward não seja necessário
-		checkCmd := exec.Command("curl", "-s", "--max-time", "1", "-o", "/dev/null", "-w", "%{http_code}", "http://localhost:8000")
-		var out bytes.Buffer
-		checkCmd.Stdout = &out
-
-		if checkCmd.Run() != nil || !strings.Contains(strings.TrimSpace(out.String()), "200") {
-			fmt.Println("\n⚠️ Detectado problema na conexão com o frontend.")
-			fmt.Println("   Reconfigurando port-forwards para garantir acesso...")
-
-			// Forçar reconfiguração de port-forwards
-			err := setupPortForward("girus")
-			if err != nil {
-				fmt.Println("   ⚠️", err)
-				fmt.Println("   Configure manualmente: kubectl port-forward -n girus svc/girus-frontend 8000:80 --address 0.0.0.0")
-			} else {
-				fmt.Println("   ✅ Port-forwards reconfigurados com sucesso!")
-			}
-		}
-	}
-
-	// Desenhar uma linha separadora
-	fmt.Println("\n" + strings.Repeat("─", 60))
-
-	// Exibir informações sobre o laboratório adicionado
-	fmt.Println("✅ LABORATÓRIO ADICIONADO COM SUCESSO!")
-
-	if labTitle != "" && labID != "" {
-		fmt.Printf("\n📚 Título: %s\n", labTitle)
-		fmt.Printf("🏷️  ID: %s\n", labID)
-	} else if labID != "" {
-		fmt.Printf("\n🏷️  ID do Laboratório: %s\n", labID)
-	}
-
-	fmt.Println("\n📋 PRÓXIMOS PASSOS:")
-	fmt.Println("  • Acesse o Girus no navegador para usar o novo laboratório:")
-	fmt.Println("    http://localhost:8000")
-
-	fmt.Println("\n  • Para ver todos os laboratórios disponíveis via CLI:")
-	fmt.Println("    girus list labs")
-
-	fmt.Println("\n  • Para verificar detalhes do template adicionado:")
-	if labID != "" {
-		fmt.Printf("    kubectl describe configmap -n girus | grep -A20 %s\n", labID)
-	} else {
-		fmt.Println("    kubectl get configmaps -n girus -l app=girus-lab-template")
-		fmt.Println("    kubectl describe configmap <nome-do-configmap> -n girus")
-	}
-
-	// Linha final
-	fmt.Println(strings.Repeat("─", 60))
-}
-
-// checkPortForwardNeeded verifica se o port-forward para o backend precisa ser reconfigurado
-func checkPortForwardNeeded() bool {
-	backendNeeded := false
-	frontendNeeded := false
-
-	// Verificar se a porta 8080 (backend) está em uso
-	backendPortCmd := exec.Command("lsof", "-i", ":8080")
-	if backendPortCmd.Run() != nil {
-		// Porta 8080 não está em uso, precisamos de port-forward
-		backendNeeded = true
-	} else {
-		// Porta está em uso, mas precisamos verificar se é o kubectl port-forward e se está funcional
-		// Verificar se o processo é kubectl port-forward
-		backendProcessCmd := exec.Command("sh", "-c", "ps -eo pid,cmd | grep 'kubectl port-forward' | grep '8080' | grep -v grep")
-		if backendProcessCmd.Run() != nil {
-			// Não encontrou processo de port-forward ativo ou válido
-			backendNeeded = true
-		} else {
-			// Verificar se a conexão com o backend está funcionando
-			backendHealthCmd := exec.Command("curl", "-s", "--head", "--max-time", "2", "http://localhost:8080/api/v1/health")
-			backendNeeded = backendHealthCmd.Run() != nil // Retorna true (precisa de port-forward) se o comando falhar
-		}
-	}
-
-	// Verificar se a porta 8000 (frontend) está em uso
-	frontendPortCmd := exec.Command("lsof", "-i", ":8000")
-	if frontendPortCmd.Run() != nil {
-		// Porta 8000 não está em uso, precisamos de port-forward
-		frontendNeeded = true
-	} else {
-		// Porta está em uso, mas precisamos verificar se é o kubectl port-forward e se está funcional
-		// Verificar se o processo é kubectl port-forward
-		frontendProcessCmd := exec.Command("sh", "-c", "ps -eo pid,cmd | grep 'kubectl port-forward' | grep '8000' | grep -v grep")
-		if frontendProcessCmd.Run() != nil {
-			// Não encontrou processo de port-forward ativo ou válido
-			frontendNeeded = true
-		} else {
-			// Verificar se a conexão com o frontend está funcionando
-			frontendCheckCmd := exec.Command("curl", "-s", "--max-time", "2", "-o", "/dev/null", "-w", "%{http_code}", "http://localhost:8000")
-			var out bytes.Buffer
-			frontendCheckCmd.Stdout = &out
-			if frontendCheckCmd.Run() != nil {
-				frontendNeeded = true
-			} else {
-				statusCode := strings.TrimSpace(out.String())
-				frontendNeeded = !(statusCode == "200" || statusCode == "301" || statusCode == "302")
-			}
-		}
-	}
-
-	// Se qualquer um dos serviços precisar de port-forward, retorne true
-	return backendNeeded || frontendNeeded
 }
 
 func init() {
